Moved function to stops.py and fixed merge issues
- Made preprocess.py match main
- Adjusted imports for clean_stops
